Updated to c++20
- Updated to c++20
- Setup now installs premake version 5.0.0-beta2 (required for c++20)
- updated EnTT
- fixed all warnings
- Removed depricated UUID::IsValid
- Removed Utils
- Changed Utils::BytesToString to String::BytesToString

+++ Shark/src/Platform/DirectX11/DirectXPipeline.cpp
@@ -129,7 +129,7 @@ namespace Shark {
 			inputElements.emplace_back(inputElementDesc);
 		}
 		const auto& shaderBinaryWithInputSignature = m_Specification.Shader.As<DirectXShader>()->GetShaderBinaries().at(ShaderUtils::ShaderStage::Vertex);
-		SK_DX11_CALL(dev->CreateInputLayout(inputElements.data(), inputElements.size(), shaderBinaryWithInputSignature.data(), shaderBinaryWithInputSignature.size(), &m_InputLayout));
+		SK_DX11_CALL(dev->CreateInputLayout(inputElements.data(), (UINT)inputElements.size(), shaderBinaryWithInputSignature.data(), shaderBinaryWithInputSignature.size(), &m_InputLayout));
 
 	}
